dev_requirements.txt
django
django-autoslug
django-debug-toolbar
django-localflavor
django-model-utils
django-grappelli
djangorestframework
factory-boy
psycopg2
pytest
pytest-cov
pytest-django